chore(infra): wire transmit, refine geoip, update swagger, deps and migrations
- Adonis Transmit added: config/transmit.ts, start/transmit.ts, registered in
  adonisrc.ts and start/events.ts.
- New shared service ProviderErrorReporter to centralize provider error logging.
- Geoip service and middleware tightened (lookup error handling).
- Shared exceptions handler adjusted alongside the new transmit/listener wiring.
- Swagger config and admin user details DTO touch-ups for the new admin views.
- Migrations: users and kyc_level table tweaks aligned with the kyc/auth changes.
- Dependency updates in package.json / package-lock.json; docker-compose.yml
  bumped for the local stack.

+++ database/migrations/17298456241_create_users_table.ts
@@ -6,11 +6,12 @@ export default class extends BaseSchema {
   async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
-      table
-        .integer('country_id')
+     table
+        .bigInteger('country_id')
         .unsigned()
         .nullable()
         .references('id')
+
         .inTable('countries')
         .onDelete('SET NULL')
         .onUpdate('CASCADE')
